+++ examples/model/0007_analyzers.py
@@ -69,10 +69,10 @@ def main():
     for beam in model.beams:
         viewer.scene.add(beam.centerline)
         viewer.scene.add(beam.geometry)
-        viewer.scene.add(Tag(text=str(beam.graph_node), position=beam.centerline.midpoint, height=40, color=Color.black()))
+        viewer.scene.add(Tag(text=str(beam.graphnode), position=beam.centerline.midpoint, height=40, color=Color.black()))
 
     tol = Tolerance(absolute=0.01)
-    num_of_beams = len(list(model.beams))
+    num_of_beams = len(model.beams)
 
     ### find triplets only
     # analyzer = TripletAnalyzer(model, tolerance=Tolerance(absolute=0.01))